pyproject.toml
[project]
name = "wagent"
version = "0.1.0"
description = "Web版ChatGPTをAPIとして利用するためのブリッジツール"
authors = [
    { name = "nezumi0627" }
]
readme = "README.md"
requires-python = ">= 3.10"
license = { text = "MIT" }
keywords = ["chatgpt", "automation", "playwright", "api", "bridge"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    
    # Browser Automation
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.6",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "requests>=2.32.0",
    
    # Optional: For cookie management
    "browser-cookie3>=0.19.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
wagent = "wagent.main:main"

[project.urls]
Homepage = "https://github.com/nezumi0627/wagent"
Repository = "https://github.com/nezumi0627/wagent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["wagent"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_ignores = true
